问题描述:pwm波不能修改脉宽 问题原因:电机调速接口和pwm接口入参不一致 解决办法:更改入参并修复定时器分频

+++ Driver/bsp/inc/motor.h
@@ -3,9 +3,9 @@
 
 void Motor_Init(void);
 void Motor_Astate(uint8_t state);
-void Motor_Aspeed(int8_t speed);
+void Motor_Aspeed(uint16_t speed);
 void Motor_Bstate(uint8_t state);
-void Motor_Bspeed(int8_t speed);
+void Motor_Bspeed(uint16_t speed);
 unsigned char show1(void);
 unsigned char show2(void);
